requirements.txt
async-generator==1.10
attrs==22.1.0
browsermob-proxy==0.8.0
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
MouseInfo==0.1.3
numpy==1.23.5
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
Pillow==9.3.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dotenv==0.21.0
python3-xlib==0.15
pytweening==1.0.4
requests==2.28.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
webdriver-manager==3.8.5
wsproto==1.2.0
